feat(frontend): add link components
