+++ apps/frontend/src/components/PageTitle.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { Flex, Heading, IconButton } from '@chakra-ui/react';
+import { Link } from 'react-router-dom';
+import { HStack, Heading, IconButton } from '@chakra-ui/react';
 import { ChevronLeftIcon } from '@chakra-ui/icons';
 
 type PageTitleProps = {
@@ -8,16 +9,18 @@ type PageTitleProps = {
 
 function PageTitle({ title }: PageTitleProps) {
   return (
-    <Flex w={'100%'} align={'center'} mb={10} justify={'start'}>
-      <IconButton
-        colorScheme=""
-        aria-label="Back"
-        icon={<ChevronLeftIcon as="button" w={8} h={8} color="brand.500" />}
-      />
+    <HStack w={'100%'} mb={10}>
+      <Link to="..">
+        <IconButton
+          colorScheme="transparent"
+          aria-label="Back"
+          icon={<ChevronLeftIcon w={8} h={8} color="brand.500" />}
+        />
+      </Link>
       <Heading size={'sm'} m={'auto'}>
         {title}
       </Heading>
-    </Flex>
+    </HStack>
   );
 }
